+++ src/graph/depth_first_search.rs
@@ -3,14 +3,14 @@
 use super::Graph;
 
 pub struct DepthFirstSearch {
-    marked: Vec<bool>, // 是否已访问过。
-    edge_to: Vec<i32>, // 起点到某顶点的最后一个顶点。
-    s: i32,            // 起点。
+    marked: Vec<bool>,   // 是否已访问过。
+    edge_to: Vec<usize>, // 起点到某顶点的最后一个顶点。
+    s: usize,            // 起点。
 }
 
 impl DepthFirstSearch {
     /// 创建一个图的深度优先结果。
-    pub fn new(g: Graph, s: i32) -> Self {
+    pub fn new(g: Graph, s: usize) -> Self {
         let marked = vec![false; g.v()];
         let edge_to = vec![0; g.v()];
         let mut d = Self { marked, edge_to, s };
@@ -19,23 +19,23 @@ impl DepthFirstSearch {
     }
 
     /// 深度优先算法。
-    fn dfs(&mut self, g: &Graph, s: i32) {
-        self.marked[s as usize] = true;
+    fn dfs(&mut self, g: &Graph, s: usize) {
+        self.marked[s] = true;
         for w in g.adj(s) {
-            if !self.marked[*w as usize] {
-                self.edge_to[*w as usize] = s;
+            if !self.marked[*w] {
+                self.edge_to[*w] = s;
                 self.dfs(g, *w);
             }
         }
     }
 
     /// 是否存在 s 到 v 的路径。
-    pub fn has_path_to(&self, w: i32) -> bool {
-        self.marked[w as usize]
+    pub fn has_path_to(&self, w: usize) -> bool {
+        self.marked[w]
     }
 
     /// 获取 s 到 v 的路径。
-    pub fn path_to(&self, v: i32) -> Vec<i32> {
+    pub fn path_to(&self, v: usize) -> Vec<usize> {
         if !self.has_path_to(v) {
             return vec![];
         }
@@ -44,7 +44,7 @@ impl DepthFirstSearch {
         let mut x = v;
         while x != self.s {
             path.push(x);
-            x = self.edge_to[x as usize];
+            x = self.edge_to[x];
         }
         path.push(self.s);
         path
@@ -53,8 +53,8 @@ impl DepthFirstSearch {
 
 #[cfg(test)]
 mod tests {
-    use crate::graph::graph_data;
     use super::*;
+    use crate::graph::graph_data;
 
     #[test]
     fn has_path_to_test() {
